bbblack - lz4 - patch the lz4 file
This patch makes lz4 compile-able. It also adds a memcpy function.
The function is needed because the bootloader is running in
no MMU mode. In no MMU mode, the memory space is treated as
strongly-ordered. An unaligned access on strongly ordered memory
leads to an assignment fault. To avoid this, a simple inefficient
byte-copy routine is implemented.